requirements.txt
openai
python-telegram-bot
# python3==3.9.7
django